+++ core/model/src/main/java/com/example/store/core/model/Order.kt
@@ -4,9 +4,9 @@ import com.example.store.core.model.cart.DeliveryMethod
 
 data class Order(
     val id: Int = 0,
-    val cartTotal: Double = 0.0,
-    val deliveryFee: Double = 0.0,
-    val orderTotal: Double = 0.0,
+    val cartTotal: Int = 0,
+    val deliveryFee: Int = 0,
+    val orderTotal: Int = 0,
     val deliveryLocationName: String = "DEFAULT_LOCATION_NAME",
     val latitude: Double = 0.0,
     val longitude: Double = 0.0,
